+++ packages/backends/hive-gateway/src/server.ts
@@ -6,8 +6,28 @@ import { readFileSync } from 'fs';
 import { createServer, Server as HttpServer } from 'http';
 import { Server } from 'socket.io';
 import express from 'express';
+import passport from 'passport';
+import neo4j from 'neo4j-driver'
 
+var OidcStrategy = require('passport-openidconnect').Strategy;
 const greenlock = require("greenlock-express")
+
+
+const {NODE_ENV} = process.env
+
+
+
+const url = process.env.AUTH_SERVER || "auth.hexhive.io"
+const config = {
+	issuer: `https://${url}`,
+	authorizationURL: `https://${url}/auth`,
+	tokenURL: `https://${url}/token`,
+	userInfoURL: `https://${url}/me`,
+	clientID: process.env.CLIENT_ID || "test" || `${NODE_ENV != "production" ? "staging-" : ""}hexhive.io`,
+	clientSecret: process.env.CLIENT_SECRET || `${NODE_ENV != "production" ? "staging-" : ""}hexhive_secret`,
+	callbackURL: `${process.env.BASE_URL || "http://localhost:7000"}/callback`,
+	scope: process.env.SCOPE || "openid email name groups"
+};
 
 const argv = yargs(hideBin(process.argv)).options({
 	port: {type: 'number', default: 7000},
@@ -19,7 +39,83 @@ const argv = yargs(hideBin(process.argv)).options({
 
 	const app = express()
 	const server = createServer(app)
+
+	const neoDriver = neo4j.driver(
+		process.env.NEO4J_URI || "localhost",
+		neo4j.auth.basic(process.env.NEO4J_USER || "neo4j", process.env.NEO4J_PASSWORD || "test")
+	)
 	
+	app.use(passport.initialize())
+	app.use(passport.session())
+
+	passport.serializeUser((user, next) => {
+		console.log("serializeUser", user)
+		next(null, user);
+	});
+	  
+	passport.deserializeUser((obj: any, next) => {
+		console.log("deserializeUser", obj);
+		
+		next(null, obj)
+		// next(null, {...obj, name: "Test"});
+	});
+
+	app.use('/login', (req, res, next) => {
+		if(req.query.returnTo){
+			(req.session as any).returnTo = req.query.returnTo
+		}
+		next();
+	}, passport.authenticate('oidc'))
+
+	app.get('/logout', function(req, res){
+		req.logout();
+		res.redirect('/');
+	});
+
+	app.use('/callback',
+		passport.authenticate('oidc', { failureRedirect: '/error' }),
+			(req, res) => {
+
+				const returnTo = (req.session as any).returnTo;
+				(req.session as any).returnTo = undefined;
+				console.log(req)
+				res.redirect(returnTo || process.env.UI_URL || "https://next.hexhive.io/dashboard");
+			}
+		);
+
+		
+	passport.use('oidc', new OidcStrategy({
+		...config,
+		skipUserProfile: false
+	}, (issuer: any, profile: any, done: any) => {
+		console.log({profile})
+		const session = neoDriver?.session();
+		session?.run(`
+		  MATCH (org:HiveOrganisation)-[:TRUSTS]->(user:HiveUser {id: $id})
+		  CALL {
+			  WITH user
+			MATCH (user)-[:HAS_ROLE]->()-->(apps:HiveAppliance)
+			RETURN distinct(apps{.*}) as apps
+		  }
+		  RETURN user{
+			id: user.id,
+			name: user.name,
+			organisation: org.id,
+			applications: collect(apps{.*})
+		  }
+		`, {
+		  
+			id: profile.id,
+		  
+		}).then((data) => {
+		  
+		  const user = data.records?.[0].get(0);
+		  console.log("deserializeUser", user);
+		  session.close()
+		  done(null, user);
+		})
+	}))
+
 	const { port, endpoints, dev } = await argv.argv;
 
 	console.log(`=> Starting Gateway on ${port}`)
