+++ packages/core-apps/hive-dashboard/src/root.component.tsx
@@ -4,8 +4,10 @@ import { AuthProvider } from "@hexhive/auth-ui";
 
 const { NODE_ENV, REACT_APP_API, REACT_APP_URL } = process.env;
 
+const API_URL = localStorage.getItem('HEXHIVE_API');
+
 const client = new ApolloClient({
-  uri: REACT_APP_API ?  `${REACT_APP_API}/graphql` : 'http://localhost:7000/graphql',
+  uri: API_URL ?  `${API_URL}/graphql` : 'http://localhost:7000/graphql',
   cache: new InMemoryCache(),
   credentials: 'include'
 })
@@ -13,8 +15,8 @@ const client = new ApolloClient({
 
 export default function Root(props) {
   return <ApolloProvider client={client}>
-        <AuthProvider
-      authorizationServer={NODE_ENV == 'production' ? (REACT_APP_API || "https://staging-api.hexhive.io") : 'http://localhost:7000'}
+    <AuthProvider
+      authorizationServer={NODE_ENV == 'production' ? (API_URL || "https://staging-api.hexhive.io") : 'http://localhost:7000'}
       >
         <Dashboard />
       </AuthProvider>
