+++ packages/core/hexhive-microfrontend/src/templates/index.ejs
@@ -153,6 +153,7 @@
     You need to enable JavaScript to run this app.
   </noscript>
   <script>
+    localStorage.setItem('HEXHIVE_API', '<%= apiUrl %>');
     System.import('@hexive/root-config');
   </script>
   <import-map-overrides-full show-when-local-storage="devtools" dev-libs></import-map-overrides-full>
