+++ packages/core/hexhive-ui/src/components/GraphGrid/GraphGrid.tsx
@@ -7,8 +7,8 @@ import RGL, {
 } from "react-grid-layout";
 import { GraphContainer } from "./GraphContainer";
 
-import "/node_modules/react-grid-layout/css/styles.css";
-import "/node_modules/react-resizable/css/styles.css";
+import "react-grid-layout/css/styles.css";
+import "react-resizable/css/styles.css";
 
 // const ReactGridLayout = WidthProvider(RGL);
 const ResponsiveGridLayout = WidthProvider(Responsive);
